BottomNavigationBarを別ファイルに移動
